backend/requirements.txt
fastapi
uvicorn
pandas
scikit-learn
openai
websockets
python-dotenv
pydantic
